Homework9-oop-fixes
