+++ Homework9/src/IShape.ts.ts
@@ -1,0 +1,4 @@
+export interface IShape {
+    calcArea(): number;
+    calcPerimeter(): number;
+}
